added a check to see if there is at least one day chosen

+++ src/Commands/moderation/reset-times.js
@@ -123,6 +123,23 @@ module.exports = {
             });
             return;
         }
+
+        if (interaction.options.getBoolean('monday')) daysArray.push('monday');
+        if (interaction.options.getBoolean('tuesday')) daysArray.push('tuesday');
+        if (interaction.options.getBoolean('wednesday')) daysArray.push('wednesday');
+        if (interaction.options.getBoolean('thursday')) daysArray.push('thursday');
+        if (interaction.options.getBoolean('friday')) daysArray.push('friday');
+        if (interaction.options.getBoolean('saturday')) daysArray.push('saturday');
+        if (interaction.options.getBoolean('sunday')) daysArray.push('sunday');
+
+        // Check if the daysArray is empty
+        if (daysArray.length === 0) {
+            await interaction.editReply({
+                content: 'You must at least have one day chosen.',
+                ephemeral: true
+            });
+            return; // Exit the command execution if no days were chosen
+        }
         
         try{
             await deleteAllTeamTimesMessages(gameName, teamName, Client)
@@ -226,15 +243,6 @@ module.exports = {
         for (hour = fromHour; hour <= toHour; hour++){
             timesArray.push(hour)
         }
-
-        if (interaction.options.getBoolean('monday')) daysArray.push('monday');
-        if (interaction.options.getBoolean('tuesday')) daysArray.push('tuesday');
-        if (interaction.options.getBoolean('wednesday')) daysArray.push('wednesday');
-        if (interaction.options.getBoolean('thursday')) daysArray.push('thursday');
-        if (interaction.options.getBoolean('friday')) daysArray.push('friday');
-        if (interaction.options.getBoolean('saturday')) daysArray.push('saturday');
-        if (interaction.options.getBoolean('sunday')) daysArray.push('sunday');
-
 
         for (let day of daysArray) {
             for (let hour of timesArray) {
